wip - lint

+++ full-service/src/json_rpc/v1/models/txo.rs
@@ -255,7 +255,7 @@ mod tests {
         let txo_details = db::models::Txo::get(&txo_hex, &wallet_db.get_conn().unwrap())
             .expect("Could not get Txo");
         let txo_status = txo_details.status(&wallet_db.get_conn().unwrap()).unwrap();
-        assert_eq!(txo_details.value as u64, 15_625_000 * MOB as u64);
+        assert_eq!(txo_details.value as u64, 15_625_000 * MOB);
         let json_txo = Txo::new(&txo_details, &txo_status);
         assert_eq!(json_txo.value_pmob, "15625000000000000000");
     }
